requirements.txt
dependency-injector
pyyaml
fastapi
uvicorn

pytest
requests
pytest-cov
httpx
pytest-asyncio
aioredis
fakeredis

bitarray
aiohttp

google-cloud-firestore
mock-firestore-async
aioresponses
ruff